renamed the autoabstractor b/c it lives on its own now
